+++ HISTORY.rst
@@ -9,6 +9,15 @@
   :class:`~skosprovider.skos.Concept`. Every source is an object that currently
   only has one attribute, a citation. This looks like a good universal common
   denominator. (#20)
+- Add sorting to :meth:`skosprovider.providers.VocabularyProvider.get_all`,
+  :meth:`skosprovider.providers.VocabularyProvider.find`,
+  :meth:`skosprovider.providers.VocabularyProvider.get_top_concepts`,
+  :meth:`skosprovider.providers.VocabularyProvider.get_top_display`,
+  :meth:`skosprovider.providers.VocabularyProvider.get_children_display`.
+  Sorting can be done on `id`, `label` or `sortlabel`. The last option makes it
+  possible to introduce arbitrary sorting for concepts, eg. to sort periods
+  chronologically. The sort order can be specified with the `sort_order`
+  parameter. (#21)
 - Remove :func:`skosprovider.providers.VocabularyProvider.expand_concept` that
   was deprecated since `0.2.0`.
 - Fixed a bug with :func:`skosprovider.skos.dict_to_label` and